Support for merging string literals
In the standard, this is not actually part of the grammar, and handled
just beforeparsing. But it is correct that this can be handled by the
parser.

Since this is a small change, we add support for this in our
parsers. But we do not provide a test for this. The reason is that we
intend tests to be some kind of reference test suite for C parsing,
and string merging is not officially part of the parser.